# toop-connector/src/main/resources/openAPI/toop.yaml
openapi: 3.0.2
info:
  title: TOOP
  version: '2.1.0'
servers:
  - url: 'https://toop-connector.brreg.no/'
paths:
  /log:
    get:
      description: Return the most recent log messages
      operationId: getLog
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Log"
  /query:
    get:
      description: Query for valid country codes
      operationId: getCountryCodes
      parameters:
        - name: type
          in: query
          description: The TOOP type (GBM, eProcurement)
          required: true
          schema:
            $ref: "#/components/schemas/QueryType"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CountryCode"
  /query/LP/{countrycode}/{legalperson}:
    get:
      description: Query for a given Legal Person
      operationId: getByLegalPerson
      parameters:
        - name: countrycode
          in: path
          required: true
          description: The orgno country code
          schema:
            type: string
        - name: legalperson
          in: path
          required: true
          description: ID for the legal person
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enhet"
  /query/NP/{countrycode}/{naturalperson}:
    get:
      description: Query for a given Natural Person
      operationId: getByNaturalPerson
      parameters:
        - name: countrycode
          in: path
          required: true
          description: The orgno country code
          schema:
            type: string
        - name: naturalperson
          in: path
          required: true
          description: ID for the natural person
          schema:
            type: string
        - name: firstname
          in: query
          required: false
          description: Firstname for the natural person
          schema:
            type: string
        - name: lastname
          in: query
          required: false
          description: Lastname for the natural person
          schema:
            type: string
        - name: birthdate
          in: query
          required: false
          description: Birthdate for the natural person
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enhet"
components:
  schemas:
    Log:
      type: object
      properties:
        time:
          type: string
          format: date-time
        severity:
          type: string
        message:
          type: string
    Enhet:
      type: object
      properties:
        organisasjonsnummer:
          type: string
        navn:
          type: string
        forretningsadresse:
          $ref: "#/components/schemas/Adresse"
        postadresse:
          $ref: "#/components/schemas/Adresse"
        organisasjonsform:
          $ref: "#/components/schemas/Organisasjonsform"
        registreringsdatoEnhetsregisteret:
          type: string
        registrertIMvaregisteret:
          type: boolean
        naeringskode1:
          $ref: "#/components/schemas/Naeringskode"
        naeringskode2:
          $ref: "#/components/schemas/Naeringskode"
        naeringskode3:
          $ref: "#/components/schemas/Naeringskode"
        antallAnsatte:
          type: integer
        stiftelsedato:
          type: string
        institusjonellSektorkode:
          $ref: "#/components/schemas/Sektorkode"
        registrertIForetaksregisteret:
          type: boolean
        registrertIStiftelsesregisteret:
          type: boolean
        registrertIFrivillighetsregisteret:
          type: boolean
        sisteInnsendteAarsregnskap:
          type: string
        konkurs:
          type: boolean
        underAvvikling:
          type: boolean
        underTvangsavviklingEllerTvangsopplosning:
          type: boolean
        maalform:
          type: string
        _links:
          $ref: "#/components/schemas/Self"
    Adresse:
      type: object
      properties:
        landkode:
          type: string
        land:
          type: string
        kommunenummer:
          type: string
        kommune:
          type: string
        postnummer:
          type: string
        poststed:
          type: string
        adresse:
          type: array
          items:
            type: string
    Naeringskode:
      type: object
      properties:
        kode:
          type: string
        beskrivelse:
          type: string
    Sektorkode:
      type: object
      properties:
        kode:
          type: string
        beskrivelse:
          type: string
    Organisasjonsformer:
      type: object
      properties:
        organisasjonsformer:
          type: array
          items:
            $ref: "#/components/schemas/Organisasjonsform"
    Organisasjonsform:
      type: object
      properties:
        kode:
          type: string
        beskrivelse:
          type: string
        _links:
          $ref: "#/components/schemas/Self"
    Self:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/Href"
    Href:
      type: object
      properties:
        href:
          type: string
    CountryCode:
      type: object
      properties:
          id:
            type: string
          code:
            type: string
          name:
            type: string
    CountryCodeResult:
      type: object
      properties:
        total-result-count:
          type: integer
        used-result-count:
          type: integer
        result-page-index:
          type: integer
        result-page-count:
          type: integer
        first-result-index:
          type: integer
        last-result-index:
          type: integer
        matches:
          type: array
          items:
            $ref: "#/components/schemas/CountryCodeMatch"
    CountryCodeMatch:
      type: object
      properties:
        participantID:
          $ref: "#/components/schemas/CountryCodeParticipant"
        entities:
          type: array
          items:
            $ref: "#/components/schemas/CountryCodeParticipantEntity"
    CountryCodeParticipant:
      type: object
      properties:
        scheme:
          type: string
        value:
          type: string
    CountryCodeParticipantEntity:
      type: object
      properties:
        name:
          type: array
          items:
            $ref: "#/components/schemas/CountryCodeParticipantEntityName"
        countryCode:
          type: string
    CountryCodeParticipantEntityName:
      type: object
      properties:
        name:
          type: string
    QueryType:
      type: string
      enum:
        - GBM
        - eProcurement